Fix README highlighting, headers and structure
Fixes the syntax highlighting of the code snippets.

Fixes the relative sizing of section headers in the README. Previously,
"Installing", "Examples" and "Contributing" were structurally
represented as subsections of "Router API". Now, those sections are all
second-level headers, and the subsections of "API" are third-level
headers.

Moves the "Installation" section to the top, above "API". It seems this
is common of most modern libraries.

Changes and adds the following links:

* The page.js link at the top now links to their own GitHub repo,
  because it's more informative than the Meteor-package repo.

* Adds a link to Meteor. Overkill? Maybe, but it might help some people.

* Fixes the link to Meteorite, which previously linked to
  `possibilities/meteorite`.

* Fixes the link to page.js' Context object to point to the start of the
  Context documentation.

Generally cleans up the Markdown syntax.

Slightly rewords a few sentences.